Develop

+++ src/app/code/community/Shopgate/Cloudapi/sql/shopgate_cloudapi_setup/mysql4-upgrade-3.0.0-3.1.0.php
@@ -71,10 +71,12 @@ $table = $installer->getConnection()
 
 try {
     $installer->getConnection()->createTable($table);
-    $installer->getConnection()->query(
-        'INSERT INTO `shopgate_oauth_authorization_codes` (`authorization_code`, `client_id`, `user_id`, `expires`, `resource_id`, `resource_type`) SELECT `token`, 0, `user_id`, `expires`, `resource_id`, `type` FROM `shopgate_resource_auth_codes`;'
-    );
-    $installer->getConnection()->dropTable('shopgate_resource_auth_codes');
+    if ($installer->getConnection()->isTableExists('shopgate_resource_auth_codes')) {
+        $installer->getConnection()->query(
+            'INSERT INTO `shopgate_oauth_authorization_codes` (`authorization_code`, `client_id`, `user_id`, `expires`, `resource_id`, `resource_type`) SELECT `token`, 0, `user_id`, `expires`, `resource_id`, `type` FROM `shopgate_resource_auth_codes`;'
+        );
+        $installer->getConnection()->dropTable( 'shopgate_resource_auth_codes' );
+    }
 } catch (Exception $e) {
     Mage::logException($e);
 }
